+++ doc/source/releasenotes.rst
@@ -7,6 +7,8 @@ Release notes
 1.0.3
 ======
 * Getting thermodynamic quantities in the SGC MC now also retrieves averages from observers.
+* Added `interactive` option to :py:func:`~clease.plot_post_process.plot_eci`
+* Added :meth:`~clease.settings.ClusterExpansionSettings.get_cluster_corresponding_to_cf_name`.
 
 1.0.2
 ======